add test_condor_parameters.py
