+++ htcondor/run.py
@@ -31,6 +31,9 @@ if __name__ == '__main__':
     condor_pars= CondorParameters("parameters.yaml")
     outdir = '/'.join((condor_pars["base_outputdir"],condor_pars["subdirectory"]))
 
+    filename= "start_"+job+".txt"
+    os.system("touch "+ filename)
+
     #create the gaudi run command from the run parameters
     gaudi_command = 'LD_PRELOAD=$FCCSWBASEDIR/build.$BINARY_TAG/lib/libPapasUtils.so $FCCSWBASEDIR/run  fccrun.py {}  --rpythiainput  {} --routput output.root  --rmaxevents {}'.format(                                                                                                                                                    condor_pars["script"], condor_pars["input"], int(condor_pars["nevents"]))
     print "start gaudi run:-"
@@ -43,6 +46,7 @@ if __name__ == '__main__':
     move_command = 'xrdcp output.root {}/{}/output_{}.root'.format( condor_pars["base_outputdir"], condor_pars["subdirectory"], job)
     print "move files:- ", move_command
     os.system(move_command)
+    os.remove(filename)
     
     print "finish gaudi run"
