+++ test/test_analysis_test_cfg.py
@@ -11,8 +11,8 @@ if context.name == 'fcc':
     from heppy.test.plot_ee_b import Plotter
     from heppy.framework.looper import Looper
     from ROOT import TFile
-    if 'FCCDATASETS' in os.environ:
-        del os.environ['FCCDATASETS']                
+    if 'FCCDATASETBASEOUT' in os.environ:
+        del os.environ['FCCDATASETBASEOUT']                
     from analysis_test_cfg import config
     
     import logging
@@ -23,7 +23,7 @@ if context.name == 'fcc':
     class TestAnalysis(unittest.TestCase):
 
         def setUp(self):
-            os.environ['FCCDATASETS'] = os.path.abspath('test')                    
+            os.environ['FCCDATASETBASEOUT'] = os.path.abspath('test')                    
             random.seed(0xdeadbeef)
             self.outdir = tempfile.mkdtemp()
             import logging
