+++ test_basedir.py
@@ -7,16 +7,16 @@ import fcc_datasets
 class TestBaseDir(unittest.TestCase):
     
     def test_1(self):
-        os.environ['FCCDATASETS'] = 'blah'
+        os.environ['FCCDATASETBASEOUT'] = 'blah'
         self.assertEqual(basedir.basename(),
                          'blah')        
-        del os.environ['FCCDATASETS']
+        del os.environ['FCCDATASETBASEOUT']
         self.assertEqual(basedir.basename(),
                          '/'.join([fcc_datasets.__path__[0], 'test']))
         basedir.set_basename('foo')
         self.assertEqual(basedir.basename(),
                          'foo')
-        os.environ['FCCDATASETS'] = 'bar'       
+        os.environ['FCCDATASETBASEOUT'] = 'bar'       
         self.assertEqual(basedir.basename(),
                          'bar')
         import test.setbasedir
